Add dependencies between SSL certificate steps
If (for example) the key file is 0 bytes long and is unable to generate a key
the previous arrangement would fail on the CSR creation. Now it will skip the
following steps rather then failing everything and it can be manually
investigated laster.

In the example above delete the empty key and
--start-at 'Install DKIM senders domain lookup file' will generate it next time
around.